Create file of links to static GTFS feeds
Create list of URL strings that can be imported by other data pipeline
procedures as constants. None of these require an API key. All of them
contain the relevant files for a GTFS specification for an urban area's
transit system or some component of it, which can be read in together
and processed using the gtfs_kit or gtfs_functions libraries.