SUCCEED AT IMPORTING TRANSIT STATION DATA FOR THE FIRST TIME!!
Create ingest_transit_stations() function that takes an already-
downloaded DataFrame of stops, converts them row-by-row into
Django-model-compatible objects, and saves them to database.

Add try-except blocks so that process skips over an observation
instead of terminating when it'd attempt to save a duplicate of
data that already exists in the db.

run() is currently set to "test" this by downloading the Metra
GTFS dataset, isolating the stops dataframe, and ingesting just that