Get full transit stop and route ingestion pipeline working
When run from the command line, the file extract_scheduled_gtfs.py
should now iterate through the URLs for the static GTFS feeds for
our three cities, transform the data to prepare it for storage,
and write it to the PostGIS backend.

Two small issues to consider before merging:

1) Many stop and route names go past the 30-character limit.
When this happens, the entire upload for that stop or route is
skipped.
We may want to either consider extending that limit or truncating
names for long stops (e.g. by taking the first 27 characters and
adding "..." if the name is too long).

2) Some large bus systems seem to have multiple rows with the same
stop_id, perhaps for different directions of travel or routes that
pick up at the same stop. We should investigate further to see
if this is fine or would cause issues linking to other data.