Get routes displaying on map
Some notes:
-Routes have a tooltip that displays on-click, showing the route's
long name.
-Colors are as-is (CTA bus routes are gray, Portland bus routes are
dark bluish/purplish, New York's bus and subway routes are a wild
mixture of colors, which we might want to set to a more neutral
tone).
-Metra routes are not showing up at all (and thus have no color);
investigate whether this is a data ingestion issue, a validation
issue (such as the color column needing .strip() to be 6 characters
long for hex code), or something else such as a non-standard
column name.